+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 dev:
   - provide better error message on context deadlline exceeded
+  - update launchpad output to match latest version
 
 1.35.2:
   - update dependencies